feat: Implement template creation service with architecture-specific utilities
- Added architecture-utils.ts for managing architecture-specific component descriptions, required services, and cost estimates.
- Introduced bicep-generators.ts to create Bicep templates and parameters for different architectures.
- Developed file-generators.ts to generate various template files including azure.yaml, README.md, and devcontainer configurations.
- Created template-structures.ts to define directory and file structures for different template architectures.
- Implemented template-search.ts to list and search templates using the Azure Developer CLI.
- Added template-validation.ts for validating template structure, content, and compliance with required standards.
- Enhanced types/index.ts with new interfaces for template analysis, validation results, and creation parameters.
- Updated validation utilities in validation.ts to check for required README sections and security notices.
