+++ package.json
@@ -9,7 +9,7 @@
   },
   "files": [
     "dist"
-  ],  
+  ],
   "scripts": {
     "build": "tsc",
     "start": "node dist/index.js",
@@ -42,6 +42,7 @@
   "devDependencies": {
     "@modelcontextprotocol/sdk": "^1.8.0",
     "@types/jest": "^29.5.14",
+    "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.13.14",
     "jest": "^29.7.0",
     "rimraf": "^5.0.1",
